+++ drivers/keyboard.h
@@ -6,7 +6,7 @@
 
 void register_kbd_driver(void);
 void kbd_buffer_push(uint8_t c);
-int32_t kbd_read(struct file *f, void *top, uint32_t count);
+ssize_t kbd_read(struct file *f, void *top, size_t count);
 bool kbd_hasdata(void);
 
 #endif
